Struct miniconf::MqttClient

pub struct MqttClient<'buf, Settings, Stack, Clock, Broker, const Y: usize>
where
    Settings: TreeKey<Y> + Clone,
    Stack: TcpClientStack,
    Clock: Clock,
    Broker: Broker,
{ /* private fields */ }

MQTT settings interface.

Design

The MQTT client places the TreeKey paths <path> at the MQTT <prefix>/settings/<path> topic, where <prefix> is provided in the client constructor.

It publishes its alive-ness as a 1 to <prefix>/alive and sets a will to publish 0 there when it is disconnected.

Limitations

The MQTT client logs failures to subscribe to the settings topic, but does not re-attempt to connect to it when errors occur.

The client only supports paths up to MAX_TOPIC_LENGTH = 128 byte length. Re-publication timeout is fixed to REPUBLISH_TIMEOUT_SECONDS = 2 seconds.

Example

use miniconf::{MqttClient, Tree};

#[derive(Tree, Clone, Default)]
struct Settings {
    foo: bool,
}

let mut buffer = [0u8; 1024];
let localhost: minimq::embedded_nal::IpAddr = "127.0.0.1".parse().unwrap();
let mut client: MqttClient<'_, _, _, _, minimq::broker::IpBroker, 1> = MqttClient::new(
    std_embedded_nal::Stack::default(),
    "quartiq/application/12345", // prefix
    std_embedded_time::StandardClock::default(),
    Settings::default(),
    minimq::ConfigBuilder::new(localhost.into(), &mut buffer).keepalive_interval(60),
)
.unwrap();

client
    .handled_update(|path, old_settings, new_settings| {
        if new_settings.foo {
            return Err("Foo!");
        }
        *old_settings = new_settings.clone();
        Ok(())
    })
    .unwrap();

Implementations

impl<'buf, Settings, Stack, Clock, Broker, const Y: usize> MqttClient<'buf, Settings, Stack, Clock, Broker, Y>
where for<'de> Settings: JsonCoreSlash<'de, Y> + Clone, Stack: TcpClientStack, Clock: Clock + Clone, Broker: Broker,

pub fn new( stack: Stack, prefix: &str, clock: Clock, settings: Settings, config: ConfigBuilder<'buf, Broker> ) -> Result<Self, ProtocolError>

Construct a new MQTT settings interface.

Args

stack - The network stack to use for communication.

prefix - The MQTT device prefix to use for this device.

clock - The clock for managing the MQTT connection.

settings - The initial settings values.

config - The configuration of the MQTT client.

pub fn handled_update<F, E>( &mut self, handler: F ) -> Result<bool, Error<Stack::Error>>
where F: FnMut(&str, &mut Settings, &Settings) -> Result<(), E>, E: Display,

Update the MQTT interface and service the network. Pass any settings changes to the handler supplied.

Args

handler - A closure called with updated settings that can be used to apply current settings or validate the configuration. Arguments are (path, old_settings, new_settings).

Returns

True if the settings changed. False otherwise.

pub fn update(&mut self) -> Result<bool, Error<Stack::Error>>

Update the settings from the network stack without any specific handling.

Returns

True if the settings changed. False otherwise

pub fn settings(&self) -> &Settings

Get the current settings from miniconf.

pub fn force_republish(&mut self)

Force republication of the current settings.

Note

This is intended to be used if modification of a setting had side effects that affected another setting.