miniconf
Struct Path

#[repr(transparent)]
pub struct Path<T: ?Sized, const S: char>(pub T);

Path with named keys separated by a separator char

The path will either be empty or start with the separator.

path: T: A Write to write the separators and node names into during Transcode. See also TreeKey::traverse_all() and Metadata::max_length() for upper bounds on path length. Can also be a AsRef<str> to implement IntoKeys (see KeysIter).

const S: char: The path hierarchy separator to be inserted before each name, e.g. '/'.

Tuple Fields
0: T
Implementations

impl<T: ?Sized, const S: char> Path<T, S>

pub const fn separator(&self) -> char

The path hierarchy separator

impl<T, const S: char> Path<T, S>

pub fn into_inner(self) -> T

Extract just the path