Trait miniconf::JsonCoreSlash

pub trait JsonCoreSlash<'de, const Y: usize = 1>: TreeSerialize<Y> + TreeDeserialize<'de, Y> {
    // Required methods
    fn set_json(
        &mut self,
        path: &str,
        data: &'de [u8]
    ) -> Result<usize, Error<Error>>;
    fn get_json(
        &self,
        path: &str,
        data: &mut [u8]
    ) -> Result<usize, Error<Error>>;
    fn set_json_by_index(
        &mut self,
        indices: &[usize],
        data: &'de [u8]
    ) -> Result<usize, Error<Error>>;
    fn get_json_by_index(
        &self,
        indices: &[usize],
        data: &mut [u8]
    ) -> Result<usize, Error<Error>>;
}

Miniconf with “JSON and /”.

Access items with '/' as path separator and JSON (from serde-json-core) as serialization/deserialization payload format.

Required Methods

fn set_json( &mut self, path: &str, data: &'de [u8] ) -> Result<usize, Error<Error>>

Update an element by path.

Args

path - The path to the element. Everything before the first '/' is ignored.

data - The serialized data making up the content.

Returns

The number of bytes consumed from data or an Error.

fn get_json(&self, path: &str, data: &mut [u8]) -> Result<usize, Error<Error>>

Retrieve a serialized value by path.

Args

path - The path to the element.

data - The buffer to serialize the data into.

Returns

The number of bytes used in the data buffer or an Error.

fn set_json_by_index( &mut self, indices: &[usize], data: &'de [u8] ) -> Result<usize, Error<Error>>

Update an element by indices.

Args

indices - The indices to the element. Everything before the first '/' is ignored.

data - The serialized data making up the content.

Returns

The number of bytes consumed from data or an Error.

fn get_json_by_index( &self, indices: &[usize], data: &mut [u8] ) -> Result<usize, Error<Error>>

Retrieve a serialized value by indices.

Args

indices - The indices to the element.

data - The buffer to serialize the data into.

Returns

The number of bytes used in the data buffer or an Error.

Object Safety

This trait is not object safe.