Trait miniconf::TreeKey

pub trait TreeKey<const Y: usize = 1> {
    // Required methods
    fn name_to_index(name: &str) -> Option<usize>;
    fn traverse_by_key<K, F, E>(keys: K, func: F) -> Result<usize, Error<E>>
       where K: Iterator,
             K::Item: Key,
             F: FnMut(usize, &str) -> Result<(), E>;
    fn metadata() -> Metadata;

    // Provided methods
    fn path<K, P>(keys: K, path: P, sep: &str) -> Result<usize, Error<Error>>
       where K: IntoIterator,
             K::Item: Key,
             P: Write { ... }
    fn indices<'a, K, I>(
        keys: K,
        indices: I
    ) -> Result<usize, Error<SliceShort>>
       where K: IntoIterator,
             K::Item: Key,
             I: IntoIterator<Item = &'a mut usize> { ... }
    fn iter_paths<P: Write>(sep: &str) -> PathIter<'_, Self, Y, P>  { ... }
    fn iter_paths_unchecked<P: Write>(sep: &str) -> PathIter<'_, Self, Y, P>  { ... }
}

Traversal, iteration, and serialization/deserialization of nodes in a tree.

The following documentation sections on TreeKey<Y> apply analogously to TreeSerialize<Y> and TreeDeserialize<Y>.

Recursion

The TreeKey trait (and the TreeSerialize/TreeDeserialize traits as well) are meant to be implemented recursively on nested data structures. Recursion here means that a container that implements TreeKey, may call on the TreeKey implementations of inner types.

The const parameter Y in the traits here is the recursion depth and determines the maximum nesting of TreeKey layers. It’s at least 1 and defaults to 1.

The recursion depth Y doubles as an upper bound to the key length (the depth/height of the tree): An implementor of TreeKey<Y> may consume at most Y items from the keys iterator argument in the recursive methods (TreeSerialize::serialize_by_key(), TreeDeserialize::deserialize_by_key(), TreeKey::traverse_by_key()). This includes both the items consumed directly before recursing and those consumed indirectly by recursing into inner types. In the same way it may call func in TreeKey::traverse_by_key() at most Y times, again including those calls due to recursion into inner Miniconf types.

This implies that if an implementor T of TreeKey<Y> (with Y >= 1) contains and recurses into an inner type using that type’s TreeKey<Z> implementation, then 1 <= Z <= Y must hold and T may consume at most Y - Z items from the keys iterators and call func at most Y - Z times.

Keys

The keys used to locate nodes can be either iterators over usize or iterators over &str names.

usize may appear like ASN.1 Object Identifiers. &str keys are sequences of names, like path names. When concatenated, they are separated by path hierarchy separators, e.g. '/'.

Derive macros

Derive macros to automatically implement the correct TreeKey<Y> traits on a struct S are available through crate::TreeKey, crate::TreeSerialize, and crate::TreeDeserialize. A shorthand derive macro that derives all three trait implementations is also available at crate::Tree.

To derive TreeSerialize/TreeDeserialize, each field in the struct must either implement serde::Serialize/serde::de::DeserializeOwned (and ultimately also be supported by the intended serde::Serializer/serde::Deserializer backend) or implement the respective TreeSerialize/TreeDeserialize trait themselves for the required remaining recursion depth.

For each field, the remaining recursion depth is configured through the #[tree(depth(Y))] attribute, with Y = 1 being the implied default when using #[tree()] and Y = 0 invalid. If the attribute is not present, the field is a leaf and accessed only through its serde::Serialize/serde::Deserialize implementation. With the attribute present the field is accessed through its TreeKey<Y> implementation with the given remaining recursion depth.

Array

Blanket implementations of the TreeKey traits are provided for homogeneous arrays [T; N] up to recursion depth Y = 8.

When a [T; N] is used as TreeKey<Y> (i.e. marked as #[tree(depth(Y))] in a struct) and Y > 1 each item of the array is accessed as a TreeKey tree. For a depth Y = 0 (attribute absent), the entire array is accessed as one atomic value. For Y = 1 each index of the array is is instead accessed as one atomic value.

The type to use depends on the desired semantics of the data contained in the array. If the array contains TreeKey items, one can (and often wants to) use Y >= 2. However, if each element in the array should be individually configurable as a single value (e.g. a list of u32), then Y = 1 can be used. With Y = 0 all items are to be accessed simultaneously and atomically. For e.g. [[T; 2]; 3] where T: TreeKey<3> the recursion depth is Y = 5. It automatically implements TreeKey<5>. For [[T; 2]; 3] where T: Serialize + DeserializeOwned, any Y <= 2 is available.

Option

Blanket implementations of the TreeKey traits are provided for Option<T> up to recursion depth Y = 8.

These implementation do not alter the path hierarchy and do not consume any items from the keys iterators. The TreeKey behavior of an Option is such that the None variant makes the corresponding part of the tree inaccessible at run-time. It will still be iterated over by TreeKey::iter_paths() but attempts to TreeSerialize::serialize_by_key() or TreeDeserialize::deserialize_by_key() them return Error::Absent. This is intended as a mechanism to provide run-time construction of the namespace. In some cases, run-time detection may indicate that some component is not present. In this case, namespaces will not be exposed for it.

If the depth specified by the #[tree(depth(Y))] attribute exceeds 1, the Option can be used to access within the inner type using its TreeKey trait. If there is no tree attribute on an Option field in a struct or in an array, JSON nullcorresponds toNoneas usual and theTreeKey` trait is not used.

The following example shows potential usage of arrays and Option:

#[derive(TreeKey)]
struct S {
    // "/b/1/2" = 5
    #[tree(depth(2))]
    b: [[u32; 3]; 3],
    // "/c/0" = [3,4], optionally absent at runtime
    #[tree(depth(2))]
    c: [Option<[u32; 2]>; 2],
}

Generics

The macros add bounds to generic types of the struct they are acting on. If a generic type parameter T of the struct S<T>is used as a type parameter to a field type a: F1<F2<T>> the type T will be considered to reside at type depth X = 2 (as it is within F2 which is within F1) and the following bounds will be applied:

With the #[tree()] attribute not present on a, T will receive bounds Serialize/DeserializeOwned when TreeSerialize/TreeDeserialize is derived.

With #[tree(depth(Y))], and Y - X < 1 it will also receive bounds Serialize + DeserializeOwned.

For Y - X >= 1 it will receive the bound T: TreeKey<Y - X>.

E.g. In the following T resides at depth 2 and T: TreeKey<1> will be inferred:

#[derive(TreeKey)]
struct S<T> {
    #[tree(depth(3))]
    a: [Option<T>; 2],
};
// This works as [u32; N] implements TreeKey<1>:
S::<[u32; 5]>::metadata();
// This does not compile as u32 does not implement TreeKey<1>:
// S::<u32>::metadata();

This behavior is upheld by and compatible with all implementations in this crate. It is only violated when deriving TreeKey for a struct that (a) forwards its own type parameters as type parameters to its field types, (b) uses TreeKey on those fields, and (c) those field types use their type parameters at other levels than TreeKey<Y - 1>. See the test_derive_macro_bound_failure test in tests/generics.rs.

Example

See the crate documentation for an example showing how the traits and the derive macros work.

Required Methods

fn name_to_index(name: &str) -> Option<usize>

Convert a node name to a node index.

The details of the mapping and the usize index values are an implementation detail and only need to be stable for at runtime.

#[derive(TreeKey)]
struct S {
    foo: u32,
    bar: u16,
}
assert_eq!(S::name_to_index("bar"), Some(1));

fn traverse_by_key<K, F, E>(keys: K, func: F) -> Result<usize, Error<E>>
where K: Iterator, K::Item: Key, F: FnMut(usize, &str) -> Result<(), E>,

Call a function for each node on the path described by keys.

Traversal is aborted once func returns an Err(E).

May not exhaust keys if a leaf is found early. i.e. keys may be longer than required. If Self is a leaf, nothing will be consumed from keys and Ok(0) will be returned. If Self is non-leaf (internal node) and the iterator is exhausted (empty), Err(Error::TooShort(0)) will be returned.

#[derive(TreeKey)]
struct S {
    foo: u32,
    bar: u16,
};
assert_eq!(
    S::traverse_by_key(["bar"].into_iter(), |index, name| {
        assert_eq!((1, "bar"), (index, name));
        Ok::<_, ()>(())
    }),
    Ok(1)
);

Args

keys: An Iterator of Keys identifying the node.

func: A FnMut to be called for each node on the path. Its arguments are the index and the name of the node at the given depth. Returning Err() aborts the traversal.

Returns

Final node depth on success

fn metadata() -> Metadata

Get metadata about the paths in the namespace.

#[derive(TreeKey)]
struct S {
    foo: u32,
    bar: u16,
};
let m = S::metadata();
assert_eq!((m.max_depth, m.max_length, m.count), (1, 3, 2));

Provided Methods

fn path<K, P>(keys: K, path: P, sep: &str) -> Result<usize, Error<Error>>
where K: IntoIterator, K::Item: Key, P: Write,

Convert keys to path.

This is typically called through a PathIter returned by TreeKey::iter_paths.

keys may be longer than required. Extra items are ignored.

#[derive(TreeKey)]
struct S {
    foo: u32,
    bar: u16,
};
let mut s = String::new();
S::path([1], &mut s, "/").unwrap();
assert_eq!(s, "/bar");

Args

keys: An Iterator of Keys identifying the node.

path: A string to write the separators and node names into. See also TreeKey::metadata() for upper bounds on path length.

sep: The path hierarchy separator to be inserted before each name.

Returns

Final node depth on success

fn indices<'a, K, I>(keys: K, indices: I) -> Result<usize, Error<SliceShort>>
where K: IntoIterator, K::Item: Key, I: IntoIterator<Item = &'a mut usize>,

Convert keys to indices.

This determines the indices of the item specified by keys.

See also TreeKey::path() for the analogous function.

Entries in indices at and beyond the depth returned are unaffected.

#[derive(TreeKey)]
struct S {
    foo: u32,
    bar: u16,
};
let mut i = [0; 2];
let depth = S::indices(["bar"], &mut i).unwrap();
assert_eq!(&i[..depth], &[1]);

Args

keys: An Iterator of Keys identifying the node.

indices: An iterator of &mut usize to write the node indices into. If indices is shorter than the node depth, Error<SliceShort> is returned See also TreeKey::metadata() for upper bounds on depth.

Returns

Final node depth on success

fn iter_paths<P: Write>(sep: &str) -> PathIter<'_, Self, Y, P>

Create an iterator of all possible paths.

This is a depth-first walk. The iterator will walk all paths, including those that may be absent at run-time (see Option). The iterator has an exact and trusted Iterator::size_hint.

#[derive(TreeKey)]
struct S {
    foo: u32,
    bar: u16,
};
let paths: Vec<String> = S::iter_paths("/").map(|p| p.unwrap()).collect();
assert_eq!(paths, ["/foo", "/bar"]);

Generics

P - The type to hold the path. Needs to be core::fmt::Write + Default

Args

sep - The path hierarchy separator

Returns

An iterator of paths with a trusted and exact Iterator::size_hint().

fn iter_paths_unchecked<P: Write>(sep: &str) -> PathIter<'_, Self, Y, P>

Create an unchecked iterator of all possible paths.

See also TreeKey::iter_paths.

#[derive(TreeKey)]
struct S {
    foo: u32,
    bar: u16,
};
let paths: Vec<String> = S::iter_paths_unchecked("/").map(|p| p.unwrap()).collect();
assert_eq!(paths, ["/foo", "/bar"]);

Generics

P - The type to hold the path. Needs to be core::fmt::Write + Default.

Args

sep - The path hierarchy separator

Returns

A iterator of paths.

Object Safety

This trait is not object safe.