miniconf
Trait Transcode

pub trait Transcode {
    // Required method
    fn transcode<M, K>(&mut self, keys: K) -> Result<Node, Traversal>
       where M: TreeKey + ?Sized,
             K: IntoKeys;
}

Look up an IntoKeys in a TreeKey and transcode it.

Required Methods

fn transcode<M, K>(&mut self, keys: K) -> Result<Node, Traversal>
where M: TreeKey + ?Sized, K: IntoKeys,

Perform a node lookup of a K: IntoKeys on a M: TreeKey<Y> and transcode it.

This modifies self such that afterwards Self: IntoKeys can be used on M again. It returns a Node with node type and depth information.

Returning Err(Traversal::TooShort(depth)) indicates that there was insufficient capacity and a key could not be encoded at the given depth.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.