Trait miniconf::Transcode

pub trait Transcode {
    // Required method
    fn transcode<M, const Y: usize, K>(
        &mut self,
        keys: K,
    ) -> Result<Node, Traversal>
       where M: TreeKey<Y> + ?Sized,
             K: IntoKeys;
}

Look up an IntoKeys in a TreeKey and transcode it.

Required Methods

fn transcode<M, const Y: usize, K>( &mut self, keys: K, ) -> Result<Node, Traversal>
where M: TreeKey<Y> + ?Sized, K: IntoKeys,

Perform a node lookup of a K: IntoKeys on a M: TreeKey<Y> and transcode it.

This modifies self such that afterwards Self: IntoKeys can be used on M again. It returns a Node with node type and depth information.

Returning Err(Traversal::TooShort(depth)) indicates that there was insufficient capacity and a key could not be encoded at the given depth.

Object Safety

This trait is not object safe.