miniconf

Trait Transcode

pub trait Transcode {
    type Error;

    // Required method
    fn transcode(
        &mut self,
        schema: &Schema,
        keys: impl IntoKeys,
    ) -> Result<(), DescendError<Self::Error>>;
}

Look up an IntoKeys in a Schema and transcode it.

Required Associated Types

type Error

The possible error when transcoding.

Use this to indicate no space or unencodable/invalid values

Required Methods

fn transcode( &mut self, schema: &Schema, keys: impl IntoKeys, ) -> Result<(), DescendError<Self::Error>>

Perform a node lookup of a K: IntoKeys on a Schema and transcode it.

This modifies self such that afterwards Self: IntoKeys can be used on M again. It returns a Node with node type and depth information.

Returning Err(ValueError::Absent) indicates that there was insufficient capacity and a key could not be encoded at the given depth.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.